+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-27T09:48:13.786Z",
+    "2026-05-27T10:03:35.049Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {