+++ packages/orq-rc/RELEASES.md
@@ -9479,3 +9479,13 @@ Based on:
 - [typescript v4.10.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.45 - packages/orq-rc
+
+## 2026-05-27 10:06:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.767.0 (2.888.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.46 - packages/orq-rc